+++ src/jk_mediawiki/__init__.py
@@ -1,7 +1,7 @@
 ﻿
 
 
-__version__ = "0.2021.11.20"
+__version__ = "0.2021.12.1"
 
 
 
@@ -21,6 +21,5 @@ from .MediaWikiLocalUserInstallationMgr import MediaWikiLocalUserInstallationMgr
 from .LocalMediaWikisMgr import LocalMediaWikisMgr
 
 ################################################################################################################################
-