+++ src/jk_mediawiki/impl/WikiPHPProcessFilter.py
@@ -87,7 +87,8 @@ class WikiPHPProcessFilter(AbstractProcessFilter):
 	#
 
 	def invalidate(self):
-		self.__source.invalidate()
+		self.__filter1.invalidate()
+		self.__filter2.invalidate()
 	#
 
 #