+++ HISTORY.md
@@ -3,9 +3,13 @@
 	* code refactored
 	* improved verbose output
 	* improved dealing with errors in installed extensions
+
 * 2021-11-14
 	* added: scanning the skin directory
 	* improved: extension scanning and extension information code
 	* improved: layout of storage directories (incl. autodetect)
 	* fixed: dependencies
+
+* 2021-12-01
+	* improved: identification of processes